week6 실습
