+++ week2/vite-project/src/Input.jsx
@@ -1,7 +1,9 @@
-import { useState } from "react";
+import { useContext, useState } from "react";
 import './App.css';
+import { TodoContext } from "../context/ToDoContext";
 
-const Input = ({addTodoFunc}) => {
+const Input = () => {
+    const {addTodo} = useContext(TodoContext);
     const [formText, setFormText] = useState('');
     const onSubmit = (e) => {
         e.preventDefault();
@@ -15,7 +17,7 @@ const Input = ({addTodoFunc}) => {
             <input className="input-box" type="text" value={formText} onChange={onChange} />
             <button className="add-button"
             onClick={()=>{
-                addTodoFunc(formText)
+                addTodo(formText)
                 setFormText('')
                 }}>할 일 등록</button>
         </form>
