+++ week2/vite-project/src/main.jsx
@@ -1,8 +1,11 @@
 // import { StrictMode } from 'react'
 import { createRoot } from 'react-dom/client'
 import App from './App.jsx'
+import TodoContextProvider from '../context/ToDoContext.jsx'
 // import './index.css'
 
 createRoot(document.getElementById('root')).render(
-    <App />
+    <TodoContextProvider>
+        <App />
+    </TodoContextProvider>
 )
